refactor: integrate match adoption into existing Adopt function
